Remove float16 option, to be added again after

+++ aquilign/segmenter/models.py
@@ -8,7 +8,8 @@ import transformers
 
 def save_bert_embeddings():
 	myBertModel = transformers.BertModel.from_pretrained('google-bert/bert-base-multilingual-cased')
-	word_embeddings = myBertModel.get_input_embeddings().weight.data.half()
+	word_embeddings = myBertModel.get_input_embeddings().weight.data
+	# word_embeddings = myBertModel.get_input_embeddings().weight.data.half()
 	torch.save(word_embeddings, "aquilign/segmenter/embeddings.npy")
 
 class GRU_Encoder(nn.Module):
